% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_col_names}
\alias{extract_col_names}
\title{Title: Extract key colnames from the Denovix data frame}
\usage{
extract_col_names(xdf)
}
\arguments{
\item{xdf}{The data frame for colname(s) extraction.}
}
\value{
A vector of key column names.
}
\description{
Title: Extract key colnames from the Denovix data frame
}
\examples{
fpath <- system.file("extdata", "rnaspec2018.csv", package = "tidyDenovix", mustWork = TRUE)
rna_data = read_denovix_data(fpath, file_type = 'csv')
col_names = extract_col_names(rna_data)
}
\author{
Tingwei Adeck
}
