% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tD_subordinates.R
\name{extract_sample_names}
\alias{extract_sample_names}
\title{Title: Extract sample names from the Denovix data frame}
\usage{
extract_sample_names(dfile)
}
\arguments{
\item{dfile}{The denovix raw file for sample name(s) extraction.}
}
\value{
A vector of sample names.
}
\description{
Title: Extract sample names from the Denovix data frame
}
\examples{
fpath <- system.file("extdata", "rnaspec2018.csv", package = "tidyDenovix", mustWork = TRUE)
esn = extract_sample_names(fpath)
}
\author{
Tingwei Adeck
}
