% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tD_subordinates.R
\name{extract_wavelength}
\alias{extract_wavelength}
\title{Title: Extract wavelength}
\usage{
extract_wavelength(xdf)
}
\arguments{
\item{xdf}{The original data frame derived from importing Denovix data.}
}
\value{
A numeric data frame for the wavelength attribute.
}
\description{
Title: Extract wavelength
}
\examples{
fpath <- system.file("extdata", "rnaspec2018.csv", package = "tidyDenovix", mustWork = TRUE)
rna_data = read_denovix_data(fpath, file_type = 'csv')
wl = extract_wavelength(rna_data)
}
\author{
Tingwei Adeck
}
