% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tD_subordinates.R
\name{file_ext}
\alias{file_ext}
\title{Title: File Extension Finder}
\usage{
file_ext(epath)
}
\arguments{
\item{epath}{File path.}
}
\value{
A string representing the file extension.
}
\description{
Title: File Extension Finder
}
\examples{
fpath <- system.file("extdata", "rnaspec2018.csv", package = "tidyDenovix", mustWork = TRUE)
ext = file_ext(fpath)
}
\author{
Unknown (Adapted by Tingwei Adeck)
}
