% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_wavelength}
\alias{make_wavelength}
\title{Title: Make wavelength}
\usage{
make_wavelength()
}
\value{
A numeric data frame for the wavelength attribute.
}
\description{
Title: Make wavelength
}
\examples{
wl = make_wavelength()
}
\author{
Tingwei Adeck
}
