% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tD_subordinates.R
\name{read_denovix_data}
\alias{read_denovix_data}
\title{Title: Read Denovix files}
\usage{
read_denovix_data(dfile, file_type = c("csv", "excel", "txt"))
}
\arguments{
\item{dfile}{A Denovix file or path to the Denovix file.}

\item{file_type}{The file type being imported.}
}
\value{
A data frame.
}
\description{
A function read Denovix data files.
}
\note{
Denovix files can be saved as csv, txt or even excel files.
This function accounts for these file types.
}
\examples{
fpath <- system.file("extdata", "rnaspec2018.csv", package = "tidyDenovix", mustWork = TRUE)
rna_data = read_denovix_data(fpath, file_type = 'csv')
}
\author{
Tingwei Adeck
}
