% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tD_subordinates.R
\name{REnum}
\alias{REnum}
\title{Title: Enum function}
\usage{
REnum(elist)
}
\arguments{
\item{elist}{A list with variable binding.}
}
\value{
The list supplied by the user.
}
\description{
Title: Enum function
}
\examples{
cols = list('x260' = 1, 'x280' = 2, 'x230' = 3)
qc = REnum(cols)
a = (qc$x260)
}
\author{
Tingwei Adeck
}
