% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna1}
\alias{rna1}
\title{rna1.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 151 columns.
}
\usage{
rna1
}
\description{
An excel file version of the Denovix Spectrophotometry data.
}
\keyword{datasets}
