% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna2}
\alias{rna2}
\title{rna2.}
\format{
An object of class \code{data.frame} with 15 rows and 151 columns.
}
\usage{
rna2
}
\description{
A csv file version of the Denovix Spectrophotometry data.
}
\keyword{datasets}
