% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna3}
\alias{rna3}
\title{rna3.}
\format{
An object of class \code{data.frame} with 15 rows and 152 columns.
}
\usage{
rna3
}
\description{
A csv file version of the Denovix Spectrophotometry data.
}
\keyword{datasets}
