% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spec}
\alias{spec}
\title{spec.}
\format{
An object of class \code{data.frame} with 15 rows and 151 columns.
}
\usage{
spec
}
\description{
A txt file version of the Denovix Spectrophotometry data.
}
\keyword{datasets}
