% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDenovix-package.R
\docType{package}
\name{tidyDenovix-package}
\alias{tidyDenovix-package}
\alias{_PACKAGE}
\title{tidyDenovix: Cleans Spectrophotometry Data Obtained from the Denovix DS-11 Instrument}
\description{
Cleans spectrophotometry data obtained from the Denovix instrument. The package also provides an option to normalize the data in order to compare the quality of the samples obtained.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlphaPrime7/tidyDenovix}
  \item \url{https://alphaprime7.github.io/tidyDenovix/}
  \item Report bugs at \url{https://github.com/AlphaPrime7/tidyDenovix/issues}
}

}
\keyword{internal}
