% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tD_subordinates.R
\name{qc_attributes}
\alias{qc_attributes}
\title{Title: Quality Control data frame}
\usage{
qc_attributes(dfile, file_type = NULL, xdf)
}
\arguments{
\item{dfile}{The Denovix file path.}

\item{file_type}{The type of file.}

\item{xdf}{The Denovix data frame.}
}
\value{
A quality control data frame.
}
\description{
Title: Quality Control data frame
}
\examples{
fpath <- system.file("extdata", "rnaspec2018.csv", package = "tidyDenovix", mustWork = TRUE)
rna_data = read_denovix_data(fpath, file_type = 'csv')
qc_attributes = qc_attributes(fpath, file_type = 'csv', rna_data)
}
\author{
Tingwei Adeck
}
